requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
et-xmlfile==1.1.0
filelock==3.15.4
gdown==5.2.0
grapheme==0.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
joblib==1.4.2
numpy==2.0.0
openai==1.43.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
